python/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
idna==3.3
ImageHash==4.2.1
imageio==2.13.1
install==1.3.5
kiwisolver==1.3.2
matplotlib==3.4.3
Naked==0.1.31
networkx==2.6.3
numpy==1.21.2
opencv-contrib-python==4.5.4.60
opencv-python==4.5.3.56
packaging==21.3
Pillow==8.4.0
pyasn1==0.4.8
pycryptodome==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
pyzbar @ git+https://github.com/sushil-bharati/pyzbar.git@ba1a91416dff07dc29506026ad1aceb14f002814
qrcode==7.3.1
requests==2.26.0
scikit-image==0.19.0
scipy==1.7.1
shellescape==3.8.1
six==1.16.0
tifffile==2021.11.2
urllib3==1.26.7
setuptools~=58.4.0
